Separate input parsing from input data
